Created getWordInGrid.h file for lab 6
